+++ aggregate-info/pom.xml
@@ -93,7 +93,6 @@
             <version>4.10.0</version>
         </dependency>
     </dependencies>
-
     <build>
         <plugins>
             <plugin>
