+++ aggregate-info/src/main/java/com/xin/aggregateInfo/AggregateInfoApplication.java
@@ -13,7 +13,7 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 **/
 @EnableDiscoveryClient
 @SpringBootApplication
-@MapperScan("com.xin.aggregateInfo.mapper")
+@MapperScan("com.xin.aggregateInfo.mappers.master")
 @EnableSwagger2
 public class AggregateInfoApplication {
