--- aggregate-info/src/main/java/com/xin/aggregateInfo/mapper/JokeCollectionMapper.java
@@ -1,18 +1,0 @@
-package com.xin.aggregateInfo.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.xin.aggregateInfo.pojo.entity.JokeCollection;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- *  Mapper 接口
- * </p>
- *
- * @author xin
- * @since 2021-12-11
- */
-@Repository
-public interface JokeCollectionMapper extends BaseMapper<JokeCollection> {
-
-}
