+++ aggregate-info/src/main/java/com/xin/aggregateInfo/mappers/master/FileUploadInfoMapper.java
@@ -1,0 +1,18 @@
+package com.xin.aggregateInfo.mappers.master;
+
+import com.xin.aggregateInfo.pojo.entity.FileUploadInfo;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import org.springframework.stereotype.Repository;
+
+/**
+ * <p>
+ * 文件上传 Mapper 接口
+ * </p>
+ *
+ * @author xin
+ * @since 2022-08-20
+ */
+@Repository
+public interface FileUploadInfoMapper extends BaseMapper<FileUploadInfo> {
+
+}
