+++ aggregate-info/src/main/java/com/xin/aggregateInfo/mappers/master/JokeCollectionMapper.java
@@ -1,0 +1,18 @@
+package com.xin.aggregateInfo.mappers.master;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.xin.aggregateInfo.pojo.entity.JokeCollection;
+import org.springframework.stereotype.Repository;
+
+/**
+ * <p>
+ *  Mapper 接口
+ * </p>
+ *
+ * @author xin
+ * @since 2021-12-11
+ */
+@Repository
+public interface JokeCollectionMapper extends BaseMapper<JokeCollection> {
+
+}
