+++ aggregate-info/src/main/java/com/xin/aggregateInfo/service/CommonService.java
@@ -37,4 +37,10 @@ public interface CommonService {
      * @return
      */
     List<String> uploadBatchByMinio(List<MultipartFile> files, String dbTable);
+
+    /**
+     * pdf转word文件
+     * @param file
+     */
+    void pdfToWord(MultipartFile file,String path);
 }
