--- aggregate-info/src/main/resources/mapper/MusicInformationMapper.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.xin.aggregateInfo.mapper.MusicInformationMapper">
-
-    <select id="getList" resultType="com.xin.aggregateInfo.pojo.entity.MusicInformation">
-        select * from music_information
-        <where>
-            <if test="musicName != null and musicName != ''">
-                music_name like concat("%",#{musicName},"%")
-            </if>
-        </where>
-    </select>
-</mapper>
