+++ aggregate-info/src/main/resources/mappers/master/AggregateAdminMapper.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.xin.aggregateInfo.mapper.master.AggregateAdminMapper">
+
+    <select id="getRoleList" resultType="com.xin.vo.AggregateAdminVO$RoleVO">
+        SELECT
+            menu_id,perm
+        FROM
+            aggregate_menu
+        WHERE
+                menu_id IN (
+                SELECT
+                    menu_id
+                FROM
+                    aggregate_role_menu
+                WHERE
+                    role_id = #{roleId})
+    </select>
+</mapper>
