+++ aggregate-info/src/main/resources/mappers/master/AreaOrgCodeMapper.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.xin.aggregateInfo.mapper.master.AreaOrgCodeMapper">
+
+    <select id="getTreeList" resultType="com.xin.aggregateInfo.pojo.vo.AreaOrgCodeVO">
+        select org_code,org_name,parent_code,weight from area_org_code where status = 1 and parent_code = #{parentCode} order by weight
+    </select>
+</mapper>
