+++ aggregate-info/src/main/resources/mappers/master/MailInfoMapper.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.xin.aggregateInfo.mapper.master.MailInfoMapper">
+
+</mapper>
